refactor: refactor code as per review
