+++ src/components/bread-crumbs/bread-crumbs.tsx
@@ -1,7 +1,7 @@
 import styles from './bread-crumbs.module.css';
 import { Wrapper } from '../wrapper/wrapper';
 import { LinksBreadCrumbs } from '../links-bread-crumbs/links-bread-crumbs';
-import { usePageInfo } from '../../utils/hooks/usePageInfo';
+import { usePageInfo } from '../../utils/hooks/use-page-info';
 
 export const BreadCrumbs = () => {
   const { title } = usePageInfo();
