+++ src/pages/about-page/about-developers/developers-list.ts
